rename: 패키지명을 더 명확하게 변경및 파일 이름 수정
issue: #24
