+++ data/src/main/java/buy/coke/zet/data/api/AuthApiService.kt
@@ -1,4 +1,4 @@
-package buy.coke.zet.data.remote
+package buy.coke.zet.data.api
 
 import buy.coke.zet.data.dto.request.LoginRequestDto
 import buy.coke.zet.data.dto.request.RefreshRequestDto
@@ -9,7 +9,7 @@ import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
 
-interface AuthService {
+interface AuthApiService {
     @POST("/api/auth/login")
     suspend fun login(@Body request: LoginRequestDto): Response<CommonResponseDto<LoginResponseDto>>
